backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
pytest==8.3.4
python-docx==1.1.2
python-multipart==0.0.20
